requirements.txt
pandas
openpyxl
nibabel
matplotlib
numpy
jupyter
xlrd
scikit-image
pillow
tensorflow